Minor improvements.

+++ demos/LINUX_PC_GCC/main.c
@@ -72,7 +72,7 @@ int main(int argc, char const * const argv[])
     /* Get a new random number. */
     numbers[idx] = TbxRandomNumberGet();
     /* Print the value. */
-    printf("Random number %u: %u.\n", idx+1, (uint32_t)numbers[idx]);
+    printf("Random number %u: %u.\n", (uint32_t)idx+1, (uint32_t)numbers[idx]);
   }
 
   /* Set program exit code. */
